refactor: combine novel parsing logic into utils
