+++ lib/routes/pixiv/api/get-novels-nsfw.ts
@@ -2,10 +2,8 @@ import got from '../pixiv-got';
 import { maskHeader } from '../constants';
 import queryString from 'query-string';
 import { config } from '@/config';
-import { load } from 'cheerio';
 import { JSDOM, VirtualConsole } from 'jsdom';
 
-import getIllustDetail from './get-illust-detail';
 import pixivUtils from '../utils';
 import ConfigNotFoundError from '@/errors/types/config-not-found';
 import cache from '@/utils/cache';
@@ -165,106 +163,6 @@ async function getNovelDetail(novel_id: string, token: string): Promise<nsfwNove
     }
 }
 
-async function parseNovelContent(novelDetail: nsfwNovelDetail, token: string): Promise<string> {
-    try {
-        if (!novelDetail?.text) {
-            return '';
-        }
-
-        let content = novelDetail.text;
-
-        // 收集所有需要處理的圖片 ID 和頁碼
-        const imageMatches = [...content.matchAll(/\[pixivimage:(\d+)(?:-(\d+))?\]/g)];
-        const imageIdToUrl = new Map<string, string>();
-
-        // 批量獲取圖片資訊
-        await Promise.all(
-            imageMatches.map(async ([, illustId, pageNum]) => {
-                if (!illustId) {
-                    return;
-                }
-
-                try {
-                    const illust = (await getIllustDetail(illustId, token)).data.illust;
-                    const images = pixivUtils.getImgs(illust).map((img) => img.match(/src="([^"]+)"/)?.[1] || '');
-
-                    const imageUrl = images[Number(pageNum) || 0];
-                    if (imageUrl) {
-                        imageIdToUrl.set(pageNum ? `${illustId}-${pageNum}` : illustId, imageUrl);
-                    }
-                } catch (error) {
-                    throw new Error(`Failed to fetch illust detail for ID ${illustId}: ${error instanceof Error ? error.message : String(error)}`);
-                }
-            })
-        );
-
-        // 處理特殊標籤
-        content = content
-            // 處理作者上傳的圖片
-            // Process author uploaded images
-            .replaceAll(/\[uploadedimage:(\d+)\]/g, (match, imageId) => {
-                const originalUrl = novelDetail?.images?.[imageId]?.urls?.original;
-                if (originalUrl) {
-                    const imageUrl = originalUrl.replace('https://i.pximg.net', config.pixiv.imgProxy || '');
-                    return `<img src="${imageUrl}" alt="novel illustration ${imageId}">`;
-                }
-                return match;
-            })
-            // 處理 pixiv 圖片引用
-            // Handle pixiv image references
-            .replaceAll(/\[pixivimage:(\d+)(?:-(\d+))?\]/g, (match, illustId, pageNum) => {
-                const key = pageNum ? `${illustId}-${pageNum}` : illustId;
-                const imageUrl = imageIdToUrl.get(key);
-                return imageUrl ? `<img src="${imageUrl}" alt="pixiv illustration ${illustId}${pageNum ? ` page ${pageNum}` : ''}">` : match;
-            })
-            // 基本換行和段落
-            // Basic line breaks and paragraphs
-            .replaceAll('\n', '<br>')
-            .replaceAll(/(<br>){2,}/g, '</p><p>')
-            // ruby 標籤處理
-            // Process ruby tags
-            .replaceAll(/\[\[rb:(.*?)>(.*?)\]\]/g, '<ruby>$1<rt>$2</rt></ruby>')
-            // 連結處理
-            // Handle links
-            .replaceAll(/\[\[jumpuri:(.*?)>(.*?)\]\]/g, '<a href="$2" target="_blank" rel="noopener noreferrer">$1</a>')
-            // 頁面跳轉，但由於 [newpage] 使用 hr 分隔，沒有頁數，沒必要跳轉，所以只顯示文字
-            // Page jumps, but since [newpage] uses hr separators, without the page numbers, jumping isn't needed, so just display text
-            .replaceAll(/\[jump:(\d+)\]/g, 'Jump to page $1')
-            // 章節標題
-            // Chapter titles
-            .replaceAll(/\[chapter:(.*?)\]/g, '<h2>$1</h2>')
-            // 換頁改成分隔線
-            // Convert [newpage] tags to <hr> elements
-            .replaceAll('[newpage]', '<hr>');
-
-        // 使用 cheerio 進行最後的 HTML 清理
-        // Use cheerio for final HTML cleanup
-        const $content = load(`<article><p>${content}</p></article>`);
-
-        // 優化嵌套段落處理
-        // Optimize nested paragraph handling
-        $content('p p').each((_, elem) => {
-            const $elem = $content(elem);
-            $elem.replaceWith($elem.html() || '');
-        });
-
-        // 確保標題標籤位置正確
-        // Ensure correct heading tag placement
-        $content('p h2').each((_, elem) => {
-            const $elem = $content(elem);
-            const $parent = $elem.parent('p');
-            const html = $elem.prop('outerHTML');
-            if ($parent.length && html) {
-                $parent.replaceWith(`</p>${html}<p>`);
-            }
-        });
-
-        return $content.html() || '';
-    } catch (error) {
-        throw new Error(`Error parsing novel content: ${error instanceof Error ? error.message : String(error)}`);
-    }
-}
-
 function convertPixivProtocolExtended(caption: string): string {
     const protocolMap = new Map([
         [/pixiv:\/\/novels\/(\d+)/g, 'https://www.pixiv.net/novel/show.php?id=$1'],
@@ -323,7 +221,13 @@ export async function getR18Novels(id: string, fullContent: boolean, limit: numb
 
             try {
                 const novelDetail = await getNovelDetail(novel.id, token);
-                const content = await parseNovelContent(novelDetail, token);
+                const images = Object.fromEntries(
+                    Object.entries(novelDetail.images)
+                        .filter(([, image]) => image?.urls?.original)
+                        .map(([id, image]) => [id, image.urls.original.replace('https://i.pximg.net', config.pixiv.imgProxy || '')])
+                );
+
+                const content = await pixivUtils.parseNovelContent(novelDetail.text, images, token);
 
                 return {
                     ...baseItem,
